Allow users to set a mirror for the default SUSE registry.
This feature is only applicable for SLE15 based CaaSP installations, as those
will no longer use container-feeder, but download the docker images from a
registry instead.

By default this registry will be the official suse registry (registry.suse.com),
but the customer always has the possibility to mirror it, in which case he can
then set this mirror during installation.

This mirror will then be used a fallback, if the official registry can not be
reached.

When the user enters a https:// mirror, the certificate can be verified by
supplying a fingerprint (SHA1).